refactor(client): rework/reorganize game handler hooks
